Address review issues from PR #37
Five fixes flagged during review of the menu port:

1. UIWidgets.cpp: cast float→uint8_t in RGBA8FromVec brace-init.
   Apple Clang's -Wc++11-narrowing rejected the implicit conversion
   and the build broke at this file. Build-blocking on macOS.

2. Menu.cpp: BeginChild/EndChild guards now match. The close guard
   dropped the headerSearch qualifier, so EndChild fired without a
   matching BeginChild when useColumns=true && headerSearch=false &&
   search-active — fires when the user enables "Search In Sidebar"
   then types into a 2-column section.

3. PortMenu.cpp: re-extract confirmation moved from
   SDL_ShowSimpleMessageBox to an ImGui::BeginPopupModal. The native
   SDL modal blocked inside the in-progress ImGui frame and GPU
   command encode; the ImGui modal stays in-context.

4. scripts/new-worktree.sh, scripts/package-macos.sh,
   scripts/package-windows.ps1: comments and error messages no longer
   reference the deleted build.sh / build.ps1; point users at the
   CMake targets that replaced them.

5. CLAUDE.md: dropped the build.sh --skip-extract gotcha bullet and
   rewrote the worktree-bootstrap bullet to use cmake --target.

+++ port/gui/UIWidgets.cpp
@@ -1259,7 +1259,8 @@ ImVec4 GetRandomValue(uint64_t* state) {
 }
 
 Color_RGBA8 RGBA8FromVec(ImVec4 vec) {
-    Color_RGBA8 color = { vec.x * 255, vec.y * 255, vec.z * 255, vec.w * 255 };
+    Color_RGBA8 color = { static_cast<uint8_t>(vec.x * 255), static_cast<uint8_t>(vec.y * 255),
+                          static_cast<uint8_t>(vec.z * 255), static_cast<uint8_t>(vec.w * 255) };
     return color;
 }
